+++ optaplanner-core/src/main/java/org/optaplanner/core/config/heuristic/selector/move/generic/PillarSwapMoveSelectorConfig.java
@@ -16,7 +16,6 @@
 
 package org.optaplanner.core.config.heuristic.selector.move.generic;
 
-import java.util.Comparator;
 import java.util.List;
 
 import com.thoughtworks.xstream.annotations.XStreamAlias;
@@ -25,7 +24,6 @@ import org.optaplanner.core.config.heuristic.policy.HeuristicConfigPolicy;
 import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
 import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
 import org.optaplanner.core.config.heuristic.selector.entity.pillar.PillarSelectorConfig;
-import org.optaplanner.core.config.heuristic.selector.move.MoveSelectorConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
 import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
 import org.optaplanner.core.impl.heuristic.selector.entity.pillar.PillarSelector;
@@ -35,12 +33,8 @@ import org.optaplanner.core.impl.heuristic.selector.move.generic.PillarSwapMoveS
 import static org.apache.commons.lang3.ObjectUtils.defaultIfNull;
 
 @XStreamAlias("pillarSwapMoveSelector")
-public class PillarSwapMoveSelectorConfig extends MoveSelectorConfig<PillarSwapMoveSelectorConfig> {
+public class PillarSwapMoveSelectorConfig extends AbstractPillarMoveSelectorConfig<PillarSwapMoveSelectorConfig> {
 
-    private PillarType pillarType = null;
-    private Class<? extends Comparator> pillarOrderComparatorClass = null;
-    @XStreamAlias("pillarSelector")
-    private PillarSelectorConfig pillarSelectorConfig = null;
     @XStreamAlias("secondaryPillarSelector")
     private PillarSelectorConfig secondaryPillarSelectorConfig = null;
 
@@ -50,30 +44,6 @@ public class PillarSwapMoveSelectorConfig extends MoveSelectorConfig<PillarSwapM
 //    @XStreamConverter(value = NamedCollectionConverter.class,
 //            strings = {"variableNameInclude"}, types = {String.class}, useImplicitType = false)
     private List<String> variableNameIncludeList = null;
-
-    public PillarType getPillarType() {
-        return pillarType;
-    }
-
-    public void setPillarType(final PillarType pillarType) {
-        this.pillarType = pillarType;
-    }
-
-    public Class<? extends Comparator> getPillarOrderComparatorClass() {
-        return pillarOrderComparatorClass;
-    }
-
-    public void setPillarOrderComparatorClass(final Class<? extends Comparator> pillarOrderComparatorClass) {
-        this.pillarOrderComparatorClass = pillarOrderComparatorClass;
-    }
-
-    public PillarSelectorConfig getPillarSelectorConfig() {
-        return pillarSelectorConfig;
-    }
-
-    public void setPillarSelectorConfig(PillarSelectorConfig pillarSelectorConfig) {
-        this.pillarSelectorConfig = pillarSelectorConfig;
-    }
 
     public PillarSelectorConfig getSecondaryPillarSelectorConfig() {
         return secondaryPillarSelectorConfig;
@@ -116,7 +86,6 @@ public class PillarSwapMoveSelectorConfig extends MoveSelectorConfig<PillarSwapM
     @Override
     public void inherit(PillarSwapMoveSelectorConfig inheritedConfig) {
         super.inherit(inheritedConfig);
-        pillarSelectorConfig = ConfigUtils.inheritConfig(pillarSelectorConfig, inheritedConfig.getPillarSelectorConfig());
         secondaryPillarSelectorConfig = ConfigUtils.inheritConfig(secondaryPillarSelectorConfig, inheritedConfig.getSecondaryPillarSelectorConfig());
         variableNameIncludeList = ConfigUtils.inheritMergeableListProperty(
                 variableNameIncludeList, inheritedConfig.getVariableNameIncludeList());
